Finish Line Added, Bleacher Light Shader Added
Added the Finish Line model and shaders (please put the metal on all of the parts except for the screen. The screen mesh is where the checkered shader goes).

Added light shader for Julius's bleachers.

I also started on a hologram glitch effect wanted to use it on the finish line effect but things didn't go as planned, so I will save it for when the city track is being created.